Add files via upload
In this assignment, we were tasked to create a simple Windows Form application using C# that includes a form, labels, buttons, and textboxes. The application will allow the user to input a name, age, email, and phone number, and then display the information in a message box when the user clicks a button. You will also add additional functionality to the application over the course of the week